Fixes bug 363580 RhinoRemote debugger can not find
This fixes an issue where the Rhino Debugger resumes execution of all
threads in the event a Context.enter() and Context.exit() call is made
without a thread being created (can happen as a race condition or by
design).  EventSetImpl is simply modified to not resume all threads if
the thread ID cannot be found.

Signed-off-by: Philip Lowman <philip@yhbt.com>