Bug 481706 - Update or replace JSDT parser to support ES6
Esprima parser phase 2
Includes the following

Fixes the way problems are created

Fixes the generated problems so that we produce a
CatogorizedProblem in all cases

Implements JSDoc conversion and adds comments to the JavascriptUnit

Changes to default return type for FunctionDeclaration
to ANY from VOID.Also makes the source lenght 0 and
 start point to be the same as parent.

Removes the reserved keyword restrictions from SimpleName

Implements TypeDeclarationExpression and implements the conversion from
ClassExpression to TypeDeclarationExpression

Changes the argument for RestElementName so that
nodes such as ArrayAccess can be assigned.

Changes the types for the elements in ArrayName
because they can be Expressions. Adds a complicated
test case. Fixes the NaiveASTFlattener.

Javascript function  declarations can be statements by itself. Added a
new node type FunctionDeclarationStatement so that FunctionDeclarations
are wrapped into it.

Fixes for function ranges and namespace imports

Fixes two minor issues:
(1) propagates ranges to FunctionDeclarations when they are contained on
FunctionExpressions or FunctionStatements
(2) Fixes the incorrect value for namespace import declarations

Changes the CompilationUnitStructureVisitor to use DefaultASTVisitor.
Implements ImportDeclarations for structure.

Adds a new classpath hook to load the nashorn.jar from /lib/ext
Introduces a new bundle that exports the nashorn api.

Signed-off-by: Gorkem Ercan <gorkem.ercan@gmail.com>