Bug 486776 - Refactor JSDT Code Completion
This commit refactors the JSDT code completion for use with the DOM AST.
The implementation utilizes the completion framework already in use by
template completions, and moves away from the implementation invoked by
Openable.java. The implementation proposed in this commit is not
complete, and does not yet touch on many ES6 features.

This implementation proposes identifiers based on the file in which code
assist is invoked only. Identifiers that are proposed include
identifiers declared with the 'var' keyword, identifiers assigned with
the assignment operator ('=' - note other assignment operators not
included), recently used methods and fields, etc.

The implementation uses some workarounds for the absence of type
binding. Functions will be proposed with their corresponding parameters
available.

The tests that are used to test code assist have also been upgraded from
JUnit 3 to JUnit 4. These tests are still compliant with the mvn cli
use for JUnit 3. Many tests are skipped, including the 'edited' tests
and tests that test the usage from outside of the current file.

Change-Id: I36f8345497717de03cb875de21d844da0e204c5b
Signed-off-by: Ian Stewart-Binks <istewart@redhat.com>