Bug 488975:  Create UI to contribute new JavaScript runtimes
Initial contribution for the UI to configure JavaScript runtimes in the
preference pages.

This is accessible through JavaScript / Runtimes preference.

Changes are in both the core model and the UI model (where the majority
of changes are), plus headless tests and Node.js relative changes to
avoid breaking existing functionality.

Change-Id: I13e4b37e15a8e0bbfc475157289bc7761a257c0a
Signed-off-by: Orlando Rincón <orlandor@mx1.ibm.com>