Bug 489897 - Mark Occurrences is not working on JavaScriptEditor
This commit adds new implementation of the BindingResolver which is
based exclusively on the DOM model and thus doesn't use anything from
old 'compiler' infrastructure.

It's not complete yet, work is required for adopting/checking new ES6
constructs, handling global scope, background processing, serialization
and so on. 

Nevertheless, this code creates base for further development and fixes
the bug in question :-)

Change-Id: I4c6fde988a2390da0aadd71c045141979a79fcd6
Signed-off-by: Eugene Melekhov <emvv@mail.ru>